+++ CHANGELOG.md
@@ -2,10 +2,21 @@
 
 ## 1.20.1202.1
 
+* EXOOwaMailboxPolicy
+  * Fixed an issue trying to remove a policy;
+* TeamsMessagingPolicy
+  * Added AllowUserEditMessage property.
+* TeamsMeetingPolicy
+  * Added 'OrganizerOnly' as a support value for property
+    AutoAdmittedUsers.
+  * Temporarly removed the use of AllowAnonymousUsersToDialOut
+    since it is currently disabled on the API side.
 * EXPORT
   * Fixed an issue where an Export using the -Workloads
     parameter with a Service Principal did not export
     any resource;
+* DEPENDENCIES
+  * Upgrade AzureADPreview to version 2.0.2.129;
 
 ## 1.20.1125.1